Cleaned up the autoconf instructions for ProbABEL. As a result several files can be removed from SVN and added to the ignore list. The files that are removed (and ignored) can (and should) be generated by running autoreconf -i
This way we won't store superfluous information in SVN. For example,
changes to configure.ac will lead to changes in the ./configure
script. Since it is autoreconf that generates the ./configure script,
why store the ./configure script itself? After all, we are only
interested in the change in configure.ac.

NOTE that this only affects people building ProbABEL from SVN. If you
use 'make dist' to generate a tar.gz file, that file will have the
correct ./configure script.

Also updated the doc/CHANGES.LOG file to reflect this. Added
similar instructions to doc/INSTALL so that it now reflects our new
build system.



git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/genabel/pkg/ProbABEL@765 edb9625f-4e0d-4859-8d74-9fd3b1da38cb