Fix bug #2295. The invvar matrix was not correctly subset leading to rows and columns of zeroes in the 'unmasked' invvarmatrix (when using --mmscore). If, for example, 7 IDs had an NA for a SNP than the invvarmatrix would (correctly) be 7 rows and columns smaller. However, additionally the final seven rows and (all) the last columns would be exactly 0. It turns out that the wrong invvarmatrix (from rdata, the new object, instead of rdatain) was used when copying the rows and columns from the original invvarmatrix.
The copyright of this fix lies with the ErasmusMC, Rotterdam.

Thanks to Maarten Kooyman for spotting this! Could this also be
related to my own finding in bug #1889 (also related to the invvar
matrix)? To be checked...


The small example files also included in this commit were used for
easy checking of this (the regular mmscore_InvSigma_aj.sex.age.dat is
500x500 elements, this one is 10x10). 


git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/genabel/pkg/ProbABEL@965 edb9625f-4e0d-4859-8d74-9fd3b1da38cb