ProbABEL refactoring branch: Added check for the Eigen library to autoconf. Building using Eigen is now default. To disable use ./configure --disable-eigen
At the moment it checks for Eigen/Dense and Eigen/LU. If we include other headers as well, these should be added to the appropriate AC_CHECK_HEADERS line.

The default include path for Eigen is /usr/include/eigen3, this is where Debian/Ubuntu installs the files when installing the libeigen3-dev. To specify another path use ./configure --with-eigen-include-path=/your/path/to/eigen. 



git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/genabel/branches/ProbABEL-refactoring/ProbABEL@991 edb9625f-4e0d-4859-8d74-9fd3b1da38cb