In ProbABEL: change all remaining occurences of float to double. With this change we can finally compare the results from prob files with dose files again. This is now done with a dirty hack in gendata.cpp, where we read in the data. The data is first converted to a string (stream), which is then read back as a double using strtod() (just like we do when reading text files).
I've also set the precision of the output of beta and se_beta to 9 significant digits. We should discuss if we want to keep it this precise and/or want to switch to scientific notation for these variables. 

main.cpp: also some small code layout cleanup/added comments/etc.

reg1.cpp: cotains some commented (using and preprocessor #if statement) code in which I used the functions from the Eigen library for solving the linear system instead of running transposition and Choleski decomposition ourselves. The output from this should be checked at a later time and then replace the current way of solving the linear system.



git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/genabel/pkg/ProbABEL@1203 edb9625f-4e0d-4859-8d74-9fd3b1da38cb