Added chi^2 information to the ProbABEL output for linear regression. NOTE: for palogist and pacoxph this still needs to be fixed!!!
The chi^2 values are based on the LRT. The null model is calculated at
the beginning (this was already part of ProbABEL for a long time). In
the case of missing genotype data the null model is recalculated for
that SNP only. So for people with imputed data there should be no
difference in computation time. 

This is a bit of a rough implementation. Maybe some more work is
needed to make it better (in terms of programming style/efficiency).

Changes per file:
src/main.cpp: 
 - Some small (unrelated) changes to the way progress information is printed
 - Changed output precision of beta, se_beta, chi^2 to 6 instead of 9 digits
 - around line 700 is where the recalculation of the null model is done.
src/regdata.h, src/regdata.cpp:
 - Add a function remove_snp_from_X() that removes the genotype data
   from the design matrix. This is necessary, because in order to know
   which individuals have missing genotype data (and therefore should
   be excluded from the null estimation), we first need to have the
   genotype data in.
src/reg1.cpp:
 - At the beginning of apply_model() check if we are calculating the
   null model. if so, we don't need to apply the genotypic model at
   all. 
src/eigen_mematrix.h, src/eigen_mematrix.cpp:
 - Implement the delete_column() function. When transitioning to Eigen
   this function wasn't used anywhere in the code, so it wasn't
   carried over from the mematrix files. 
src/mematri1.h, src/mematrix.h:
 - Set the col/row number argument to const in the delete_column() and
   delete_row() functions.





git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/genabel/pkg/ProbABEL@1286 edb9625f-4e0d-4859-8d74-9fd3b1da38cb