Updated the ProbABEL tests to reflect the changes in the previous commit where chi^2 values were added to the output of palinear.
Extended tests/test_cox_ngpreds2/run_models_in_R_palinear.R (actually
that directory should be renamed now...) so that the ProbABEL output
is now compared to the R output and errors are printed when there is a
descrepancy. I've set the tolerance for the checks to 1e-5.

Copyright of this change lies with the Erasmus MC.


git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/genabel/pkg/ProbABEL@1287 edb9625f-4e0d-4859-8d74-9fd3b1da38cb