As discussed on	the mailing list (and reported by the Debian packager),	scripts that are used by the user should not have a filename extension. The user doesn't care whether we use Perl, Bash, Python, etc. to get the job done.
This commit replaces all occurences of probabel.pl with	 probabel. When installing (make install) a symbolic link is made so that for now probabel.pl continues to exist. When the script is called as probabel.pl a warning message is displayed, informing the user that (s)he should upgrade their pipelines because the .pl script will be removed in a future release. 



git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/genabel/pkg/ProbABEL@1492 edb9625f-4e0d-4859-8d74-9fd3b1da38cb