ProbABEL: intermediate commit. Most checks fail.
Maarten's profiling showed that the conversion to double when reading in genetic data resulted in a huge performance bottleneck. Since the genetic data is not stored with more than a few signicicant digits anyway it doesn't even make sense to use double precision. 

With this commit I removed the conversion to double, but now most checks fail (mostly when comparing prob vs. dose input). Needs more time to fix. 


git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/genabel/pkg/ProbABEL@1528 edb9625f-4e0d-4859-8d74-9fd3b1da38cb