First commit for adding calculation of p-values to ProbABEL. This uses the statistical distributions from Boost::Math library (http://www.boost.org/doc/libs/1_55_0/libs/math/doc/html/dist.html). By default this is enabled. To disable this run: ./configure --without-boost. If the header file is not in the default include search path, use ./configure --with-boost-include-path "/some/path/to/boost_headers".
At the moment both the columns for the chi^2 values and the p-values are printed. We may want to disable printing the chi^2 values in the future, possibly enabling them with a command line switch.

The actual calculation is performed by the function pchisq() named like its R counterpart in the main_functions_dump.cpp file. In main.cpp an extra vector is created to contain the chi^2 values in double format (next to the ostringstream that was already there). This vector of doubles (one per model) is then used when calculating the p-values.

Compilation is succesfull, but the two checks that compare output against "known-good" output files fail (because I haven't updated those files yet). Compilation and checks when running ./configure --without-boost work as expected (this should be the old (v0.4.x) behaviour).

TODO: Like before (v0.4.x series) the chi^2 is not calculated when using mmscore and the 2DF model, because that requires the 2DF Wald test, which we haven't implemented yet (main.cpp lines 440--457.)


git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/genabel/branches/ProbABEL-pvals/ProbABEL@1546 edb9625f-4e0d-4859-8d74-9fd3b1da38cb