- Remove -Wextra from the dev-options in configure.ac because it results in a lot of warnings in the filevector code, which I have no intention to fix (at least not any time soon).
- Added autoconf/automake support for the doxygen documentation. If the doxygen binary is found the doxygen documentation is built in doc/doxygen-docs/ (both in HTML and LaTeX). Note that the Makefile.am doesn't check for changes to in src/*.{h,cpp}, so if you change anything there you have to manually rerun doxygen.
Note 1: autoconf generates the doc/Doxyfile from doc/Doxyfile.in and automatically inserts the current ProbABEL version number as found in the configure.ac file, so no need to update this on a new release.
Note 2: the Doxygen documentation is not distributed in the tar.gz file generated by 'make dist' or 'make distcheck', because this kind of documentation is only useful to developers, not end-users.


git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/genabel/pkg/ProbABEL@1719 edb9625f-4e0d-4859-8d74-9fd3b1da38cb