Fix ProbABEL bug #5982: "ProbABEL's make install fails on MacOS X and FreeBSD with sed error".
It turns out that the versions of sed used by these operating systems need an explicit '-e' option added when running 'sed', because the '-i' option requires a backup filename extension to be mentioned.
I haven't tested this on MaxOS, but I have tested it on FreeBSD. Since OS X descends from BSD I hope it's fixed there as well.


git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/genabel/pkg/ProbABEL@1883 edb9625f-4e0d-4859-8d74-9fd3b1da38cb