this commit fixes several bugs in ProbABEL: - #1266: "pacoxph with no covariates"; I added a \mu of 1s to the X matrix. This doesn't affect the estimates for the SNPs and has the added benefit that the function coxph_data::coxph_data(phedata &phed, gendata &gend, const int snpnum) looks much more like it's cousin regdata::regdata(phedata &phed, gendata &gend, const int snpnum, const bool ext_is_interaction_excluded)). - #6041: "ProbABEL's Cox module reports too many errors (beta may be infinite, setting beta and se to 'NaN')  "; I added some more checks following the R Survial example code. I think the checks are now almost identical. - This also contains a small "fix" in phedata.cpp in the way the model is printed on the screen. By moving the + to a different place we don't end up with things like height ~ mu +  + snp when no covariates are present.
Implementation not related to the two bug fixes:

Adding a constant mean to the Cox PH regression I lead to "X matrix is singular" messages in ProbABEL and the R checks (luckily in both ;-)), Therefore, I had to adapt the R tests to disregard this message when related to the first variable. Similarly in the pacoxph code (in the estimate function).

Moreover, for the dominant model an error occurred in the R checks:
   <simpleError in fitter(X, Y, strats, offset, init, control, weights = weights,
        method = method, row.names(mf)): NA/NaN/Inf in foreign function call (arg 6)>
So I now catch errors in run_model_coxph.R as well and filter out the error I got.

I also added a copy of the text_cox.sh script, this time for the case without covariates.


git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/genabel/branches/ProbABEL-v0.4.4-hotfix/ProbABEL@1931 edb9625f-4e0d-4859-8d74-9fd3b1da38cb