Fix Issue #15: interactions in Cox broken
This fixes the --interactions option when using pacoxph. With the
introduction of the (mathematically useless) \mu in the Cox model (which
was introduced to fix bug 1266 (pacoxph with no covariates) on R-forge,
https://r-forge.r-project.org/tracker/index.php?func=detail&aid=1266&group_id=505&atid=2058),
the extra parts of code where we adjusted for the missing \mu were
inadvertently kept. Cox + interactions must have been broken since then.