Force the use of the C++11 standard with compiler flags
After fixing the compile issue (#31) with GCC 6.5.0 compilation failed
on older GCC versions (at least on 5.4.0). With this commit we
explicitly add -std=c++11 to the CXXFLAGS in configure.ac.

This means that the minimum GCC version is now 4.8.1, because that was
the first version that was feature-complete with the C++11 standard.