Fix incorrect line number in error message
The error message that points the reader to a mismatch between the IDs
in the genotype file and the phenotype file contained the wrong line
number. To calculate the line number from the counter a constant 2 was
added. However, there is no header line or something similar, we only
need to compensate for the fact that we start counting at 0, so adding a
constant 1 is enough.