Read/open the info file only once
After this change the info file is read only once. Before, the file was
opened and read once to determine the number of SNPs and then again to
fill the various arrays with data. Now the file is only read once and
instead of arrays we use vectors to store the information from the info
file.