Convert the allmeasured int array to a boolean vector
By converting the pre-allocated array to a vector we can (in a later
commit) simply grow the vector instead of having to first read the
phenotype file completely to determine the number of samples. The aim of
the "later commit" is to reduce the number of times a file is opened and
read.