Improve status message when reading pheno data
The status message that is printed when only the first few people are to
be used (as determined by the user on the command line) was a bit
cryptic.

This commit also removes the std:: from std::cout in this and the
complementary message because this is already specified at the top of
the file.

(This is a cherry pick of commit
2ebca4d812f8fb14d12a8e1246450c3581f1024f in the read-gzipped-genotypes
branch).